LPS-130793 headless-admin-user-impl: moves User DTO transform logic to a new class: UserAccountDTOConverter
