+++ modules/apps/headless/headless-admin-user/headless-admin-user-impl/src/main/java/com/liferay/headless/admin/user/internal/resource/v1_0/UserAccountResourceImpl.java
@@ -15,28 +15,16 @@
 package com.liferay.headless.admin.user.internal.resource.v1_0;
 
 import com.liferay.announcements.kernel.service.AnnouncementsDeliveryLocalService;
-import com.liferay.asset.kernel.model.AssetTag;
-import com.liferay.asset.kernel.service.AssetTagLocalService;
-import com.liferay.headless.admin.user.dto.v1_0.EmailAddress;
-import com.liferay.headless.admin.user.dto.v1_0.OrganizationBrief;
-import com.liferay.headless.admin.user.dto.v1_0.Phone;
-import com.liferay.headless.admin.user.dto.v1_0.PostalAddress;
-import com.liferay.headless.admin.user.dto.v1_0.RoleBrief;
-import com.liferay.headless.admin.user.dto.v1_0.SiteBrief;
 import com.liferay.headless.admin.user.dto.v1_0.UserAccount;
 import com.liferay.headless.admin.user.dto.v1_0.UserAccountContactInformation;
-import com.liferay.headless.admin.user.dto.v1_0.WebUrl;
 import com.liferay.headless.admin.user.internal.dto.v1_0.converter.OrganizationResourceDTOConverter;
+import com.liferay.headless.admin.user.internal.dto.v1_0.converter.UserAccountDTOConverter;
 import com.liferay.headless.admin.user.internal.dto.v1_0.util.CustomFieldsUtil;
-import com.liferay.headless.admin.user.internal.dto.v1_0.util.EmailAddressUtil;
-import com.liferay.headless.admin.user.internal.dto.v1_0.util.PhoneUtil;
-import com.liferay.headless.admin.user.internal.dto.v1_0.util.PostalAddressUtil;
 import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderAddressUtil;
 import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderEmailAddressUtil;
 import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderListTypeUtil;
 import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderPhoneUtil;
 import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderWebsiteUtil;
-import com.liferay.headless.admin.user.internal.dto.v1_0.util.WebUrlUtil;
 import com.liferay.headless.admin.user.internal.odata.entity.v1_0.UserAccountEntityModel;
 import com.liferay.headless.admin.user.resource.v1_0.UserAccountResource;
 import com.liferay.headless.common.spi.service.context.ServiceContextRequestUtil;
@@ -44,11 +32,10 @@ import com.liferay.petra.function.UnsafeConsumer;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Address;
 import com.liferay.portal.kernel.model.Contact;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.GroupConstants;
+import com.liferay.portal.kernel.model.EmailAddress;
 import com.liferay.portal.kernel.model.ListTypeConstants;
 import com.liferay.portal.kernel.model.Organization;
-import com.liferay.portal.kernel.model.Role;
+import com.liferay.portal.kernel.model.Phone;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.model.Website;
 import com.liferay.portal.kernel.search.BooleanClauseOccur;
@@ -62,20 +49,17 @@ import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
 import com.liferay.portal.kernel.service.ContactLocalService;
-import com.liferay.portal.kernel.service.GroupService;
-import com.liferay.portal.kernel.service.RoleService;
 import com.liferay.portal.kernel.service.UserGroupRoleLocalService;
 import com.liferay.portal.kernel.service.UserService;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.CalendarFactoryUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.odata.entity.EntityModel;
+import com.liferay.portal.vulcan.dto.converter.DTOConverterRegistry;
+import com.liferay.portal.vulcan.dto.converter.DefaultDTOConverterContext;
 import com.liferay.portal.vulcan.pagination.Page;
 import com.liferay.portal.vulcan.pagination.Pagination;
 import com.liferay.portal.vulcan.resource.EntityModelResource;
-import com.liferay.portal.vulcan.util.LocalizedMapUtil;
 import com.liferay.portal.vulcan.util.SearchUtil;
 
 import java.util.Calendar;
@@ -419,8 +403,8 @@ public class UserAccountResourceImpl
 		);
 	}
 
-	private List<com.liferay.portal.kernel.model.EmailAddress>
-		_getServiceBuilderEmailAddresses(UserAccount userAccount) {
+	private List<EmailAddress> _getServiceBuilderEmailAddresses(
+		UserAccount userAccount) {
 
 		return Optional.ofNullable(
 			userAccount.getUserAccountContactInformation()
@@ -441,9 +425,7 @@ public class UserAccountResourceImpl
 		);
 	}
 
-	private List<com.liferay.portal.kernel.model.Phone>
-		_getServiceBuilderPhones(UserAccount userAccount) {
-
+	private List<Phone> _getServiceBuilderPhones(UserAccount userAccount) {
 		return Optional.ofNullable(
 			userAccount.getUserAccountContactInformation()
 		).map(
@@ -469,18 +451,6 @@ public class UserAccountResourceImpl
 		).orElse(
 			0L
 		);
-	}
-
-	private ThemeDisplay _getThemeDisplay(Group group) {
-		return new ThemeDisplay() {
-			{
-				setPortalURL(StringPool.BLANK);
-
-				if (group != null) {
-					setSiteGroupId(group.getGroupId());
-				}
-			}
-		};
 	}
 
 	private Page<UserAccount> _getUserAccountsPage(
@@ -519,151 +489,15 @@ public class UserAccountResourceImpl
 		);
 	}
 
-	private OrganizationBrief _toOrganizationBrief(Organization organization) {
-		return new OrganizationBrief() {
-			{
-				id = organization.getOrganizationId();
-				name = organization.getName();
-			}
-		};
-	}
-
-	private RoleBrief _toRoleBrief(Role role) {
-		return new RoleBrief() {
-			{
-				id = role.getRoleId();
-				name = role.getTitle(
-					contextAcceptLanguage.getPreferredLocale());
-				name_i18n = LocalizedMapUtil.getI18nMap(
-					contextAcceptLanguage.isAcceptAllLanguages(),
-					role.getTitleMap());
-			}
-		};
-	}
-
-	private SiteBrief _toSiteBrief(Group group) {
-		return new SiteBrief() {
-			{
-				id = group.getGroupId();
-				name = group.getName(
-					contextAcceptLanguage.getPreferredLocale());
-				name_i18n = LocalizedMapUtil.getI18nMap(
-					contextAcceptLanguage.isAcceptAllLanguages(),
-					group.getNameMap());
-			}
-		};
-	}
-
 	private UserAccount _toUserAccount(User user) throws Exception {
-		Contact contact = user.getContact();
-
-		return new UserAccount() {
-			{
-				additionalName = user.getMiddleName();
-				alternateName = user.getScreenName();
-				birthDate = user.getBirthday();
-				customFields = CustomFieldsUtil.toCustomFields(
-					contextAcceptLanguage.isAcceptAllLanguages(),
-					User.class.getName(), user.getUserId(), user.getCompanyId(),
-					contextAcceptLanguage.getPreferredLocale());
-				dateCreated = user.getCreateDate();
-				dateModified = user.getModifiedDate();
-				emailAddress = user.getEmailAddress();
-				familyName = user.getLastName();
-				givenName = user.getFirstName();
-				honorificPrefix =
-					ServiceBuilderListTypeUtil.getServiceBuilderListTypeMessage(
-						contact.getPrefixId(),
-						contextAcceptLanguage.getPreferredLocale());
-				honorificSuffix =
-					ServiceBuilderListTypeUtil.getServiceBuilderListTypeMessage(
-						contact.getSuffixId(),
-						contextAcceptLanguage.getPreferredLocale());
-				id = user.getUserId();
-				jobTitle = user.getJobTitle();
-				keywords = ListUtil.toArray(
-					_assetTagLocalService.getTags(
-						User.class.getName(), user.getUserId()),
-					AssetTag.NAME_ACCESSOR);
-				name = user.getFullName();
-				organizationBriefs = transformToArray(
-					user.getOrganizations(),
-					organization -> _toOrganizationBrief(organization),
-					OrganizationBrief.class);
-				roleBriefs = transformToArray(
-					_roleService.getUserRoles(user.getUserId()),
-					role -> _toRoleBrief(role), RoleBrief.class);
-				siteBriefs = transformToArray(
-					_groupService.getGroups(
-						contextCompany.getCompanyId(),
-						GroupConstants.DEFAULT_PARENT_GROUP_ID, true),
-					group -> _toSiteBrief(group), SiteBrief.class);
-				userAccountContactInformation =
-					new UserAccountContactInformation() {
-						{
-							emailAddresses = transformToArray(
-								user.getEmailAddresses(),
-								EmailAddressUtil::toEmailAddress,
-								EmailAddress.class);
-							facebook = contact.getFacebookSn();
-							jabber = contact.getJabberSn();
-							postalAddresses = transformToArray(
-								user.getAddresses(),
-								address -> PostalAddressUtil.toPostalAddress(
-									contextAcceptLanguage.
-										isAcceptAllLanguages(),
-									address, user.getCompanyId(),
-									contextAcceptLanguage.getPreferredLocale()),
-								PostalAddress.class);
-							skype = contact.getSkypeSn();
-							sms = contact.getSmsSn();
-							telephones = transformToArray(
-								user.getPhones(), PhoneUtil::toPhone,
-								Phone.class);
-							twitter = contact.getTwitterSn();
-							webUrls = transformToArray(
-								user.getWebsites(), WebUrlUtil::toWebUrl,
-								WebUrl.class);
-						}
-					};
-
-				setDashboardURL(
-					() -> {
-						Group group = user.getGroup();
-
-						if (group == null) {
-							return null;
-						}
-
-						return group.getDisplayURL(
-							_getThemeDisplay(group), true);
-					});
-				setImage(
-					() -> {
-						if (user.getPortraitId() == 0) {
-							return null;
-						}
-
-						ThemeDisplay themeDisplay = new ThemeDisplay() {
-							{
-								setPathImage(_portal.getPathImage());
-							}
-						};
-
-						return user.getPortraitURL(themeDisplay);
-					});
-				setProfileURL(
-					() -> {
-						Group group = user.getGroup();
-
-						if (group == null) {
-							return null;
-						}
-
-						return group.getDisplayURL(_getThemeDisplay(group));
-					});
-			}
-		};
+		return _userAccountDTOConverter.toDTO(
+			new DefaultDTOConverterContext(
+				contextAcceptLanguage.isAcceptAllLanguages(),
+				Collections.emptyMap(), _dtoConverterRegistry,
+				contextHttpServletRequest, user.getUserId(),
+				contextAcceptLanguage.getPreferredLocale(), contextUriInfo,
+				contextUser),
+			user);
 	}
 
 	private static final EntityModel _entityModel =
@@ -674,22 +508,16 @@ public class UserAccountResourceImpl
 		_announcementsDeliveryLocalService;
 
 	@Reference
-	private AssetTagLocalService _assetTagLocalService;
-
-	@Reference
 	private ContactLocalService _contactLocalService;
 
 	@Reference
-	private GroupService _groupService;
+	private DTOConverterRegistry _dtoConverterRegistry;
 
 	@Reference
 	private OrganizationResourceDTOConverter _organizationResourceDTOConverter;
 
 	@Reference
-	private Portal _portal;
-
-	@Reference
-	private RoleService _roleService;
+	private UserAccountDTOConverter _userAccountDTOConverter;
 
 	@Reference
 	private UserGroupRoleLocalService _userGroupRoleLocalService;
